+++ src/Models/Tournament.ts
@@ -3,8 +3,6 @@ import {Round} from "./Round.ts";
 import {PlayerHistory} from "./PlayerHistory.ts";
 import {Tools} from "../Tools.ts";
 import {MatchResult} from "./MatchResult.ts";
-import {Tiebreaker} from "./Tiebreaker.ts";
-import {MatchResultEnum} from "./MatchResultEnum.ts";
 import {RoundFactory} from "./Factories/RoundFactory.ts";
 
 export class Tournament {
@@ -88,55 +86,5 @@ export class Tournament {
         }
 
         return allPlayerHistories;
-    }
-
-    getRanking(): Tiebreaker[] {
-        let allPlayerHistories = this.getAllPlayerHistories();
-        let playerTiebreakersDictionary: { [id: string]: Tiebreaker } =
-            Object.fromEntries(allPlayerHistories.map(ph => {
-                let statistics = ph.getStatistics();
-                return [ph.player.id, {
-                    player: ph.player,
-                    kda: statistics.getKda(),
-                    rivalNames: ph.matchResults.map(r => `${r.player.name} - ${MatchResultEnum[r.result]}`),
-                    matchPoints: statistics.getMatchPoints(),
-                    matchWinPercentage: statistics.getMatchWinPercentaje(),
-                    opponentsMatchWinPercentage: 0,
-                    binary: 0,
-                }];
-            }));
-
-        let playerTiebreakers: Tiebreaker[] = []
-        for (let ph of allPlayerHistories) {
-            let omwpSum = 0;
-            let rivalCount = 0;
-            let binary = 0;
-            for (let i = 0; i < ph.matchResults.length; i++) {
-                let rival = ph.matchResults[i];
-
-                // Lose, Lose, Win, Lose, Win -> 0b10100 -> 20
-                // Last matches weight more, supposedly not fair...
-                binary += rival.result === MatchResultEnum.Win ? Math.pow(2, i) : 0;
-
-                if (rival.player.id === Tools.byeId) continue;
-
-                omwpSum += playerTiebreakersDictionary[rival.player.id].matchWinPercentage;
-                rivalCount++;
-            }
-
-            let tiebreaker = playerTiebreakersDictionary[ph.player.id];
-            // Math.max(1, rivalCount) to prevent division by zero on only bye rival. 
-            tiebreaker.opponentsMatchWinPercentage = omwpSum / Math.max(1, rivalCount);
-
-            // Power the binary to hide simplicity...
-            let pow = Math.max(1, 7 - this.getRoundCount());
-            tiebreaker.binary = Math.pow(binary, pow);
-
-            playerTiebreakers.push(tiebreaker);
-        }
-
-        playerTiebreakers.sort(Tools.compareTiebreaker);
-
-        return playerTiebreakers;
     }
 }
