+++ src/Tools.ts
@@ -1,5 +1,4 @@
 import {Player, PlayerWithStatistics} from "./Models/Player.ts";
-import {Tiebreaker} from "./Models/Tiebreaker.ts";
 
 export abstract class Tools {
     static byeId = 'X';
@@ -44,16 +43,5 @@ export abstract class Tools {
     static deleteFromArray<T>(array: T[], element: T) {
         let index = array.indexOf(element);
         array.splice(index, 1);
-    }
-
-    static compareTiebreaker(a: Tiebreaker, b: Tiebreaker, compareName: boolean = true): number {
-        // Descending
-        return b.matchPoints - a.matchPoints
-            // Descending
-            || b.opponentsMatchWinPercentage - a.opponentsMatchWinPercentage
-            // Descending
-            || b.binary - a.binary
-            // Ascending
-            || (compareName ? a.player.name.localeCompare(b.player.name) : 0);
     }
 }
