+++ src/tournamentScript.ts
@@ -1,10 +1,10 @@
-import {continueTournament} from "./Controllers/TournamentController.ts";
+import {setupTournament} from "./Controllers/TournamentController.ts";
 import {CollapseController} from "./Controllers/CollapseController.ts";
 
 export function setupApp() {
     CollapseController.showRound();
 
-    continueTournament();
+    setupTournament();
 }
 
 setupApp();
